refactor: 테스트 오류 수정
