+++ core/src/test/java/kitchenpos/core/ApplicationTest.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package kitchenpos.core;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
