+++ core/src/test/java/kitchenpos/core/application/MenuServiceTest.java
@@ -15,8 +15,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import java.math.BigDecimal;
 import java.util.List;
 
-import static kitchenpos.TestFixtureFactory.새로운_메뉴_그룹;
-import static kitchenpos.TestFixtureFactory.새로운_상품;
+import static kitchenpos.core.TestFixtureFactory.새로운_메뉴_그룹;
+import static kitchenpos.core.TestFixtureFactory.새로운_상품;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
