+++ core/src/test/java/kitchenpos/core/application/OrderTableServiceTest.java
@@ -17,8 +17,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
 
-import static kitchenpos.TestFixtureFactory.새로운_주문;
-import static kitchenpos.TestFixtureFactory.새로운_주문_테이블;
+import static kitchenpos.core.TestFixtureFactory.새로운_주문;
+import static kitchenpos.core.TestFixtureFactory.새로운_주문_테이블;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
